I added in some useful path logic
I know it doesn't have a test yet, but I'm still a new to node. Since I needed to have it do this anyways I thought I might as well share. This simply appends enviroment variables that are probably already on peoples machines so that they are included when this script is run. This makes it a more likely this won't require any configuration on a dev's (or build) machine to use.